chore: (#150) Domain 패키지 이동
