+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/common/dto/request/ChangePasswordWebRequest.kt
@@ -1,0 +1,25 @@
+package team.comit.simtong.domain.common.dto.request
+
+import org.hibernate.validator.constraints.Length
+import team.comit.simtong.global.RegexUtil
+import javax.validation.constraints.NotBlank
+import javax.validation.constraints.Pattern
+
+/**
+ *
+ * 비밀번호 변경을 요청하는 ChangePasswordWebRequest
+ *
+ * @author Chokyunghyeon
+ * @date 2022/10/14
+ * @version 1.0.0
+ **/
+data class ChangePasswordWebRequest(
+
+    @field:NotBlank
+    val password: String,
+
+    @field:NotBlank
+    @field:Pattern(regexp = RegexUtil.SECRET_PATTERN)
+    @field:Length(min = 8, max = 20)
+    val newPassword: String
+)
