+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/common/dto/response/FindEmployeeNumberWebResponse.kt
@@ -1,0 +1,13 @@
+package team.comit.simtong.domain.common.dto.response
+
+/**
+ *
+ * 사원 번호 찾기 결과를 전송하는 FindEmployeeNumberWebResponse
+ *
+ * @author Chokyunghyeon
+ * @date 2022/09/11
+ * @version 1.0.0
+ **/
+data class FindEmployeeNumberWebResponse(
+    val employeeNumber: Int
+)
