+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/user/dto/request/SignInWebRequest.kt
@@ -1,0 +1,21 @@
+package team.comit.simtong.domain.user.dto.request
+
+import javax.validation.constraints.NotBlank
+import javax.validation.constraints.NotNull
+
+/**
+ *
+ * 로그인을 요청하는 SignInWebRequest
+ *
+ * @author kimbeomjin
+ * @date 2022/09/08
+ * @version 1.0.0
+ **/
+data class SignInWebRequest(
+
+    @field:NotNull
+    val employeeNumber: Int,
+
+    @field:NotBlank
+    val password: String
+)
